+++ src/server/run/annos_range.rs
@@ -1,4 +1,5 @@
-//! Code for `/annos/variant`.
+//! Code for `/annos/range`.
+
 use actix_web::{
     get,
     web::{self, Data, Json, Path},
@@ -8,50 +9,55 @@ use strum::IntoEnumIterator;
 
 use crate::{
     common::{keys, version},
-    server::AnnoDb,
+    server::{
+        run::fetch::{fetch_pos_protobuf, fetch_pos_tsv_json},
+        run::AnnoDb,
+    },
 };
 
-use super::error::CustomError;
-use super::fetch::{fetch_pos_protobuf, fetch_var_protobuf, fetch_var_tsv_json};
+use super::{error::CustomError, WebServerData};
 
 /// Parameters for `variant_annos::handle`.
-///
-/// Defines a variant in SPDI format with a genome release specification.
 #[serde_with::skip_serializing_none]
 #[serde_with::serde_as]
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
 #[serde(rename_all = "snake_case")]
 struct Request {
-    /// Genome release specification.
+    /// Genome release version.
     pub genome_release: String,
     /// Chromosome name.
     pub chromosome: String,
-    /// 1-based position for SPDI.
-    pub pos: u32,
-    /// Reference allele bases.
-    pub reference: String,
-    /// Alterantive allele bases.
-    pub alternative: String,
+    /// 1-based start position.
+    pub start: u32,
+    /// 1-based stop position.
+    pub stop: u32,
 }
 
-impl From<Request> for keys::Var {
-    fn from(value: Request) -> Self {
-        keys::Var {
-            chrom: value.chromosome,
-            pos: value.pos as i32,
-            reference: value.reference,
-            alternative: value.alternative,
-        }
-    }
-}
-
-impl From<Request> for keys::Pos {
-    fn from(value: Request) -> Self {
+impl Request {
+    /// Conver to start `keys::Pos`.
+    pub fn start_pos(&self) -> keys::Pos {
         keys::Pos {
-            chrom: value.chromosome,
-            pos: value.pos as i32,
+            chrom: self.chromosome.clone(),
+            pos: self.start as i32,
         }
     }
+
+    /// Conver to stop `keys::Pos`.
+    pub fn stop_pos(&self) -> keys::Pos {
+        keys::Pos {
+            chrom: self.chromosome.clone(),
+            pos: self.stop as i32,
+        }
+    }
+}
+
+/// Result for `handle`.
+#[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
+struct Result {
+    /// Version of the server code.
+    pub server_version: String,
+    /// Version of the builder code.
+    pub builder_version: String,
 }
 
 /// Result for `handle`.
@@ -67,10 +73,9 @@ struct Container {
 }
 
 /// Query for annotations for one variant.
-#[allow(clippy::option_map_unit_fn)]
-#[get("/annos/variant")]
+#[get("/annos/range")]
 async fn handle(
-    data: Data<crate::server::WebServerData>,
+    data: Data<WebServerData>,
     _path: Path<()>,
     query: web::Query<Request>,
 ) -> actix_web::Result<impl Responder, CustomError> {
@@ -92,10 +97,11 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        fetch_var_protobuf::<crate::pbs::clinvar::minimal::ExtractedVcvRecordList>(
+                        fetch_pos_protobuf::<crate::pbs::clinvar::minimal::ExtractedVcvRecordList>(
                             db,
                             anno_db.cf_name(),
-                            query.clone().into_inner().into(),
+                            query.start_pos(),
+                            query.stop_pos(),
                         )
                     })
                     .transpose()?
@@ -105,7 +111,12 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        fetch_var_tsv_json(db, anno_db.cf_name(), query.clone().into_inner().into())
+                        fetch_pos_tsv_json(
+                            db,
+                            anno_db.cf_name(),
+                            query.start_pos(),
+                            query.stop_pos(),
+                        )
                     })
                     .transpose()?
                     .map(|v| annotations.insert(anno_db, v));
@@ -114,10 +125,11 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        fetch_var_protobuf::<crate::dbsnp::pbs::Record>(
+                        fetch_pos_protobuf::<crate::dbsnp::pbs::Record>(
                             db,
                             anno_db.cf_name(),
-                            query.clone().into_inner().into(),
+                            query.start_pos(),
+                            query.stop_pos(),
                         )
                     })
                     .transpose()?
@@ -127,10 +139,11 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        fetch_var_protobuf::<crate::helixmtdb::pbs::Record>(
+                        fetch_pos_protobuf::<crate::helixmtdb::pbs::Record>(
                             db,
                             anno_db.cf_name(),
-                            query.clone().into_inner().into(),
+                            query.start_pos(),
+                            query.stop_pos(),
                         )
                     })
                     .transpose()?
@@ -140,10 +153,11 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        fetch_var_protobuf::<crate::pbs::gnomad::mtdna::Record>(
+                        fetch_pos_protobuf::<crate::pbs::gnomad::mtdna::Record>(
                             db,
                             anno_db.cf_name(),
-                            query.clone().into_inner().into(),
+                            query.start_pos(),
+                            query.stop_pos(),
                         )
                     })
                     .transpose()?
@@ -161,16 +175,18 @@ async fn handle(
                             .expect("gnomAD must have db version");
 
                         if db_version.starts_with("2.") {
-                            fetch_var_protobuf::<crate::pbs::gnomad::gnomad2::Record>(
+                            fetch_pos_protobuf::<crate::pbs::gnomad::gnomad2::Record>(
                                 db,
                                 anno_db.cf_name(),
-                                query.clone().into_inner().into(),
+                                query.start_pos(),
+                                query.stop_pos(),
                             )
                         } else if db_version.starts_with("4.") {
-                            fetch_var_protobuf::<crate::pbs::gnomad::gnomad4::Record>(
+                            fetch_pos_protobuf::<crate::pbs::gnomad::gnomad4::Record>(
                                 db,
                                 anno_db.cf_name(),
-                                query.clone().into_inner().into(),
+                                query.start_pos(),
+                                query.stop_pos(),
                             )
                         } else {
                             Err(CustomError::new(anyhow::anyhow!(
@@ -193,22 +209,25 @@ async fn handle(
                             .as_ref()
                             .expect("gnomAD must have db version");
                         if db_version.starts_with("2.") {
-                            fetch_var_protobuf::<crate::pbs::gnomad::gnomad2::Record>(
+                            fetch_pos_protobuf::<crate::pbs::gnomad::gnomad2::Record>(
                                 db,
                                 anno_db.cf_name(),
-                                query.clone().into_inner().into(),
+                                query.start_pos(),
+                                query.stop_pos(),
                             )
                         } else if db_version.starts_with("3.") {
-                            fetch_var_protobuf::<crate::pbs::gnomad::gnomad3::Record>(
+                            fetch_pos_protobuf::<crate::pbs::gnomad::gnomad3::Record>(
                                 db,
                                 anno_db.cf_name(),
-                                query.clone().into_inner().into(),
+                                query.start_pos(),
+                                query.stop_pos(),
                             )
                         } else if db_version.starts_with("4.") {
-                            fetch_var_protobuf::<crate::pbs::gnomad::gnomad4::Record>(
+                            fetch_pos_protobuf::<crate::pbs::gnomad::gnomad4::Record>(
                                 db,
                                 anno_db.cf_name(),
-                                query.clone().into_inner().into(),
+                                query.start_pos(),
+                                query.stop_pos(),
                             )
                         } else {
                             Err(CustomError::new(anyhow::anyhow!(
@@ -224,17 +243,11 @@ async fn handle(
                 data.annos[genome_release][anno_db]
                     .as_ref()
                     .map(|db| {
-                        let start: keys::Pos = query.clone().into_inner().into();
-                        let start = keys::Pos {
-                            chrom: start.chrom,
-                            pos: start.pos - 2,
-                        };
-                        let stop = query.clone().into_inner().into();
                         fetch_pos_protobuf::<crate::pbs::cons::RecordList>(
                             db,
                             anno_db.cf_name(),
-                            start,
-                            stop,
+                            query.start_pos(),
+                            query.stop_pos(),
                         )
                     })
                     .transpose()?
