+++ src/server/run/genes_clinvar.rs
@@ -1,4 +1,4 @@
-//! Code for `/genes/info`.
+//! Code for `/genes/clinvar`.
 
 use actix_web::{
     get,
@@ -7,7 +7,7 @@ use actix_web::{
 };
 use prost::Message;
 
-use crate::pbs::genes;
+use crate::pbs::clinvar::per_gene::ClinvarPerGeneRecord;
 
 use super::error::CustomError;
 use serde_with::{formats::CommaSeparator, StringWithSeparator};
@@ -28,35 +28,40 @@ struct Request {
 #[serde_with::skip_serializing_none]
 struct Container {
     // TODO: add data version
-    /// The resulting gene information.
-    pub genes: indexmap::IndexMap<String, genes::base::Record>,
+    /// The resulting per-gene ClinVar information.
+    pub genes: indexmap::IndexMap<String, ClinvarPerGeneRecord>,
 }
 
 /// Query for annotations for one variant.
 #[allow(clippy::option_map_unit_fn)]
-#[get("/genes/info")]
+#[get("/genes/clinvar")]
 async fn handle(
-    data: Data<crate::server::WebServerData>,
+    data: Data<crate::server::run::WebServerData>,
     _path: Path<()>,
     query: web::Query<Request>,
 ) -> actix_web::Result<impl Responder, CustomError> {
     let genes_db = data.genes.as_ref().ok_or(CustomError::new(anyhow::anyhow!(
         "genes database not available"
     )))?;
-    let cf_genes = genes_db
-        .db
-        .cf_handle("genes")
-        .expect("no 'genes' column family");
+    let db_clinvar = genes_db
+        .db_clinvar
+        .as_ref()
+        .ok_or(CustomError::new(anyhow::anyhow!(
+            "clinvar-genes database not available"
+        )))?;
+    let cf_genes = db_clinvar
+        .cf_handle("clinvar-genes")
+        .expect("no 'clinvar-genes' column family");
     let mut genes = indexmap::IndexMap::new();
     if let Some(hgnc_id) = query.hgnc_id.as_ref() {
         for hgnc_id in hgnc_id {
-            if let Some(raw_buf) = genes_db.db.get_cf(&cf_genes, hgnc_id).map_err(|e| {
+            if let Some(raw_buf) = db_clinvar.get_cf(&cf_genes, hgnc_id).map_err(|e| {
                 CustomError::new(anyhow::anyhow!("problem querying database: {}", e))
             })? {
-                let record =
-                    genes::base::Record::decode(std::io::Cursor::new(raw_buf)).map_err(|e| {
-                        CustomError::new(anyhow::anyhow!("problem decoding value: {}", e))
-                    })?;
+                let record = crate::pbs::clinvar::per_gene::ClinvarPerGeneRecord::decode(
+                    std::io::Cursor::new(raw_buf),
+                )
+                .map_err(|e| CustomError::new(anyhow::anyhow!("problem decoding value: {}", e)))?;
                 genes.insert(hgnc_id.to_string(), record);
             } else {
                 tracing::debug!("no such gene: {}", hgnc_id);
@@ -64,15 +69,13 @@ async fn handle(
         }
     }
 
-    let cf_meta = genes_db
-        .db
+    let cf_meta = db_clinvar
         .cf_handle("meta")
         .expect("no 'meta' column family");
-    let raw_builder_version = &genes_db
-        .db
-        .get_cf(&cf_meta, "builder-version")
+    let raw_builder_version = &db_clinvar
+        .get_cf(&cf_meta, "annonars-version")
         .map_err(|e| CustomError::new(anyhow::anyhow!("problem querying database: {}", e)))?
-        .expect("database missing 'builder-version' key?");
+        .expect("database missing 'annonars-version' key?");
     let _builder_version = std::str::from_utf8(raw_builder_version)
         .map_err(|e| CustomError::new(anyhow::anyhow!("problem decoding value: {}", e)))?
         .to_string();
